// src/views/views/src/providers/interfaces/i-view-output-file.ts
/**
 * @license
 * Copyright ContextJS All Rights Reserved.
 *
 * Use of this source code is governed by an MIT-style license that can be
 * found at https://github.com/contextjs/context/blob/main/LICENSE
 */

export interface IViewOutputFile {
    filename: string;
    code: string;
    map?: string;
}